move lua script logic outside of service
